+++ cypress/scripts/run-starter-prototype.js
@@ -17,7 +17,6 @@ const testDir = path.resolve(process.env.KIT_TEST_DIR || defaultKitPath)
   const barLocation = path.join(__dirname, '..', 'fixtures', 'plugins', 'plugin-bar')
 
   await installPlugins(testDir, [
-    '@govuk-prototype-kit/step-by-step@1',
     `"file:${fooLocation}"`,
     `"file:${barLocation}"`]
   )